+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.1.3
 
+* [Add plugin function `hideMessage`](https://github.com/siyuan-note/siyuan/issues/15485)
 * [Update Constants](https://github.com/siyuan-note/petal/pull/39)
 * [Add plugin function `IProtyleOptions.handleEmptyContent`](https://github.com/siyuan-note/siyuan/issues/15398)